mcp_servers/kubectl-ai_mcp/requirements.txt
fastmcp>=0.1.0
httpx>=0.23.0
uvicorn>=0.18.0
kubernetes>=26.1.0
python-dotenv>=0.19.0
PyYAML>=6.0